Horrid hacks to get tests functional
Starts up a whole new localdb for each test. (Does this by hard-coding
the path ot the ddb library.) Starting up the db takes 1 second
overhead for each test.

Redfine core.db client-opts to be a function so we can override it in
the tests (to point to the local db just spun up).